MAINT: Move the `np.testing` annotations to `np.testing._private.utils`
